Multifiled for input component implemented
