+++ src/__ui/input/multiple-input.module.scss
@@ -1,0 +1,6 @@
+@import "../../mixins";
+
+.MultipleInputContainer {
+  border: 1px solid;
+  margin: 2px;
+}
